now ZernikeCoefficients inherits from ModalCoefficients, not the opposite
